+++ +qclab/+qgates/SWAP.m
@@ -16,7 +16,7 @@
 classdef SWAP < qclab.qgates.QGate2
   properties (Access = protected)
     %> Qubits of this SWAP gate.
-    qubits_(1,2)  int32
+    qubits_(1,2)  int64
   end
   
   methods
