+++ test/qgates/test_qclab_qgates_CRotationX.m
@@ -2,28 +2,28 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
   methods (Test)
     function test_CRotationX(test)
       crotx = qclab.qgates.CRotationX() ;
-      test.verifyEqual( crotx.nbQubits, int32(2) );    % nbQubits
+      test.verifyEqual( crotx.nbQubits, int64(2) );    % nbQubits
       test.verifyFalse( crotx.fixed );                 % fixed
       test.verifyTrue( crotx.controlled );             % controlled
-      test.verifyEqual( crotx.control, int32(0) );     % control
-      test.verifyEqual( crotx.target, int32(1) );      % target
-      test.verifyEqual( crotx.controlState, int32(1)); % controlState
+      test.verifyEqual( crotx.control, int64(0) );     % control
+      test.verifyEqual( crotx.target, int64(1) );      % target
+      test.verifyEqual( crotx.controlState, int64(1)); % controlState
       
       % matrix
       test.verifyEqual(crotx.matrix, eye(4) );
       
       % qubit
-      test.verifyEqual( crotx.qubit, int32(0) );
+      test.verifyEqual( crotx.qubit, int64(0) );
       
       % qubits
       qubits = crotx.qubits;
       test.verifyEqual( length(qubits), 2 );
-      test.verifyEqual( qubits(1), int32(0) );
-      test.verifyEqual( qubits(2), int32(1) );
+      test.verifyEqual( qubits(1), int64(0) );
+      test.verifyEqual( qubits(2), int64(1) );
       qnew = [5, 3] ;
       crotx.setQubits( qnew );
-      test.verifyEqual( table(crotx.qubits()).Var1(1), int32(3) );
-      test.verifyEqual( table(crotx.qubits()).Var1(2), int32(5) );
+      test.verifyEqual( table(crotx.qubits()).Var1(1), int64(3) );
+      test.verifyEqual( table(crotx.qubits()).Var1(2), int64(5) );
       qnew = [0, 1];
       crotx.setQubits( qnew );
       
@@ -158,9 +158,9 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
       % ctranspose
       crotx = qclab.qgates.CRotationX(1, 2, pi/3 ) ;
       crotxp = crotx';
-      test.verifyEqual( crotxp.nbQubits, int32(2) );
-      test.verifyEqual( crotxp.control, int32(1) );
-      test.verifyEqual( crotxp.target, int32(2) );
+      test.verifyEqual( crotxp.nbQubits, int64(2) );
+      test.verifyEqual( crotxp.control, int64(1) );
+      test.verifyEqual( crotxp.target, int64(2) );
       test.verifyEqual(crotxp.matrix, crotx.matrix', 'AbsTol', eps );
       
     end
@@ -172,12 +172,12 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
       angle = qclab.QAngle( theta/2 );
       crotx = qclab.qgates.CRotationX( 1, 3, angle );
       
-      test.verifyEqual( crotx.nbQubits, int32(2) );    % nbQubits
+      test.verifyEqual( crotx.nbQubits, int64(2) );    % nbQubits
       test.verifyFalse( crotx.fixed );                 % fixed
       test.verifyTrue( crotx.controlled );             % controlled
-      test.verifyEqual( crotx.control, int32(1) );     % control
-      test.verifyEqual( crotx.target, int32(3) );      % target
-      test.verifyEqual( crotx.controlState, int32(1));        % controlState
+      test.verifyEqual( crotx.control, int64(1) );     % control
+      test.verifyEqual( crotx.target, int64(3) );      % target
+      test.verifyEqual( crotx.controlState, int64(1));        % controlState
       
       test.verifyEqual( crotx.theta, theta, 'AbsTol', eps );
       test.verifyEqual( crotx.cos, cos(theta/2), 'AbsTol', eps);
@@ -186,12 +186,12 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
       % angle, custom controlState
       crotx = qclab.qgates.CRotationX( 1, 3, angle, 0 );
       
-      test.verifyEqual( crotx.nbQubits, int32(2) );    % nbQubits
+      test.verifyEqual( crotx.nbQubits, int64(2) );    % nbQubits
       test.verifyFalse( crotx.fixed );                 % fixed
       test.verifyTrue( crotx.controlled );             % controlled
-      test.verifyEqual( crotx.control, int32(1) );     % control
-      test.verifyEqual( crotx.target, int32(3) );      % target
-      test.verifyEqual( crotx.controlState, int32(0));        % controlState
+      test.verifyEqual( crotx.control, int64(1) );     % control
+      test.verifyEqual( crotx.target, int64(3) );      % target
+      test.verifyEqual( crotx.controlState, int64(0));        % controlState
       
       test.verifyEqual( crotx.theta, theta, 'AbsTol', eps );
       test.verifyEqual( crotx.cos, cos(theta/2), 'AbsTol', eps);
@@ -200,12 +200,12 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
       % theta, default controlState
       crotx = qclab.qgates.CRotationX( 1, 3, theta );
       
-      test.verifyEqual( crotx.nbQubits, int32(2) );    % nbQubits
+      test.verifyEqual( crotx.nbQubits, int64(2) );    % nbQubits
       test.verifyFalse( crotx.fixed );                 % fixed
       test.verifyTrue( crotx.controlled );             % controlled
-      test.verifyEqual( crotx.control, int32(1) );     % control
-      test.verifyEqual( crotx.target, int32(3) );      % target
-      test.verifyEqual( crotx.controlState, int32(1));        % controlState
+      test.verifyEqual( crotx.control, int64(1) );     % control
+      test.verifyEqual( crotx.target, int64(3) );      % target
+      test.verifyEqual( crotx.controlState, int64(1));        % controlState
       
       test.verifyEqual( crotx.theta, theta, 'AbsTol', eps );
       test.verifyEqual( crotx.cos, cos(theta/2), 'AbsTol', eps);
@@ -214,12 +214,12 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
       % theta, custom controlState
       crotx = qclab.qgates.CRotationX( 1, 3, theta, 0 );
       
-      test.verifyEqual( crotx.nbQubits, int32(2) );    % nbQubits
+      test.verifyEqual( crotx.nbQubits, int64(2) );    % nbQubits
       test.verifyFalse( crotx.fixed );                 % fixed
       test.verifyTrue( crotx.controlled );             % controlled
-      test.verifyEqual( crotx.control, int32(1) );     % control
-      test.verifyEqual( crotx.target, int32(3) );      % target
-      test.verifyEqual( crotx.controlState, int32(0));        % controlState
+      test.verifyEqual( crotx.control, int64(1) );     % control
+      test.verifyEqual( crotx.target, int64(3) );      % target
+      test.verifyEqual( crotx.controlState, int64(0));        % controlState
       
       test.verifyEqual( crotx.theta, theta, 'AbsTol', eps );
       test.verifyEqual( crotx.cos, cos(theta/2), 'AbsTol', eps);
@@ -228,12 +228,12 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
       % cos, sin, default controlState
       crotx = qclab.qgates.CRotationX( 1, 3, cos(theta), sin(theta) );
       
-      test.verifyEqual( crotx.nbQubits, int32(2) );    % nbQubits
+      test.verifyEqual( crotx.nbQubits, int64(2) );    % nbQubits
       test.verifyFalse( crotx.fixed );                 % fixed
       test.verifyTrue( crotx.controlled );             % controlled
-      test.verifyEqual( crotx.control, int32(1) );     % control
-      test.verifyEqual( crotx.target, int32(3) );      % target
-      test.verifyEqual( crotx.controlState, int32(1));        % controlState
+      test.verifyEqual( crotx.control, int64(1) );     % control
+      test.verifyEqual( crotx.target, int64(3) );      % target
+      test.verifyEqual( crotx.controlState, int64(1));        % controlState
       
       test.verifyEqual( crotx.theta, 2*theta, 'AbsTol', eps );
       test.verifyEqual( crotx.cos, cos(theta), 'AbsTol', eps);
@@ -242,12 +242,12 @@ classdef test_qclab_qgates_CRotationX < matlab.unittest.TestCase
       % cos, sin, custom controlState
       crotx = qclab.qgates.CRotationX( 1, 3, cos(theta), sin(theta), 0 );
       
-      test.verifyEqual( crotx.nbQubits, int32(2) );    % nbQubits
+      test.verifyEqual( crotx.nbQubits, int64(2) );    % nbQubits
       test.verifyFalse( crotx.fixed );                 % fixed
       test.verifyTrue( crotx.controlled );             % controlled
-      test.verifyEqual( crotx.control, int32(1) );     % control
-      test.verifyEqual( crotx.target, int32(3) );      % target
-      test.verifyEqual( crotx.controlState, int32(0));        % controlState
+      test.verifyEqual( crotx.control, int64(1) );     % control
+      test.verifyEqual( crotx.target, int64(3) );      % target
+      test.verifyEqual( crotx.controlState, int64(0));        % controlState
       
       test.verifyEqual( crotx.theta, 2*theta, 'AbsTol', eps );
       test.verifyEqual( crotx.cos, cos(theta), 'AbsTol', eps);
